backend/requirements.txt
# ── LLM & Agent ──────────────────────────────────────────────
langchain==0.3.14
langchain-groq==0.2.3
langchain-community==0.3.14
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langgraph==0.2.60
langgraph-checkpoint-sqlite==2.0.4

# ── Embeddings & Vector Store ─────────────────────────────────
langchain-huggingface==0.1.2
sentence-transformers==3.3.1
faiss-cpu

# ── Backend API ───────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
httpx==0.28.1

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1
